reestructuring code
